authentication-svc/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.10

pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydantic[email]>=2.7,<3

PyJWT[crypto]==2.9.0
cryptography==43.0.1

passlib[bcrypt]==1.7.4
bcrypt==3.2.2

httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1

prometheus-fastapi-instrumentator==7.1.0

